Cleanup function binding interface and internals.
There are now only two bind macros (one for void functions, and one for  functions that return a value). Also make parameter types variadic arguments to the bind functions, so there aren't specific macros for the number of parameters. Under the hood there are still functions for each number of parameters, but this is not exposed to user code.